Fix national boundaries
Because we use data from different sources, data on different administrative levels is not perfectly aligned. Previously, data from the national level was used to mark country borders on ALL administrative levels. But particularly on the regional level, that led to visualisation artifacts where regional and national borders did not match.

Now, I use data from the particular level to visualise country borders: national on the national level, regional on the regional level, municipal on the municipal level. That makes country borders move slightly when zooming in, but at least they are consistent on a zoom level.

This does not solve the issue WITHIN adminsitrative levels, in particular in Wester Balkans countries.